# internal/docs/swagger.yaml
definitions:
  models.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      start_date:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /subscriptions/create:
    post:
      consumes:
      - application/json
      description: Create a new subscription record
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: invalid JSON or invalid fields
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create a subscription
      tags:
      - subscriptions
  /subscriptions/delete:
    delete:
      description: Delete subscription by user_id and service_name
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Service Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: missing or invalid parameters
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete a subscription
      tags:
      - subscriptions
  /subscriptions/get:
    get:
      description: Get subscription by user_id and service_name
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Service Name
        in: query
        name: service_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: missing or invalid parameters
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get a subscription
      tags:
      - subscriptions
  /subscriptions/list:
    get:
      description: List all subscriptions with pagination
      parameters:
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "400":
          description: invalid parameters
          schema:
            type: string
      summary: List subscriptions
      tags:
      - subscriptions
  /subscriptions/sum:
    get:
      description: Calculate total subscription cost over a period, filtered by user_id
        and service_name
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Service Name
        in: query
        name: service_name
        type: string
      - description: Start date YYYY-MM
        in: query
        name: start_date
        type: string
      - description: End date YYYY-MM
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Total price
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: invalid parameters
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Sum subscription prices
      tags:
      - subscriptions
  /subscriptions/update:
    put:
      consumes:
      - application/json
      description: Update subscription fields
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: invalid JSON or missing fields
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Update a subscription
      tags:
      - subscriptions
swagger: "2.0"